Fix unregister method
