+++ example/root-state.ts
@@ -1,0 +1,7 @@
+import { StateInterfaceFromModule } from '../src';
+import { doggoState } from './doggos';
+import { kittenState } from './kittens';
+
+export interface RootState extends StateInterfaceFromModule<typeof kittenState> {
+  doggoState: StateInterfaceFromModule<typeof doggoState>;
+}
