+++ example/store.ts
@@ -1,20 +1,11 @@
 import Vue from 'vue';
 import Vuex from 'vuex';
 import { registerVuexTsModules, StateInterfaceFromModule } from '../src';
-import { DoggoBreed, doggoState } from './doggos';
-import { KittenBreed, kittenState } from './kittens';
+import { doggoState } from './doggos';
+import { kittenState } from './kittens';
 
 Vue.use(Vuex);
 
 export const store = new Vuex.Store({
   plugins: [registerVuexTsModules(doggoState, kittenState)],
 });
-
-export interface RootState extends StateInterfaceFromModule<typeof kittenState> {
-  doggoState: StateInterfaceFromModule<typeof doggoState>;
-}
-
-kittenState.doggoNested.commit.addDoggo({ name: 'Rover', breed: DoggoBreed.Golden, age: 10 });
-doggoState.commit.addDoggo({ name: 'Dude', breed: DoggoBreed.Golden, age: 9 });
-doggoState.commit.addDoggo({ name: 'Aristotle', breed: DoggoBreed.Basset, age: 14 });
-kittenState.commit.addKitten({ name: 'Shadow', breed: KittenBreed.Dsh, age: 8 });
